api_gateway/requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
black==23.12.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
motor==3.3.2
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.6.1
pytest==7.4.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0
httpx==0.26.0
python-jose==3.3.0